Disable paging in PartsController and update IDE metadata
Commented out paging logic in PartsController to return all parts without pagination. Updated and added various IDE, cache, and project metadata files (e.g., .dtbcache.v2, .suo, .futdcache.v2, DocumentLayout.json, .bin files) reflecting changes in build environment, project configuration, and editor state. No source code logic changes outside of the paging adjustments.

+++ Controllers/planm/EditPlanApiController.cs
@@ -847,7 +847,7 @@ namespace PlanMatr_API.Controllers.planm
         /// </summary>
         /// <param name="planogramId">The id of the planogram to check.</param>
         /// <returns>true or false.</returns>
-        private async Task<bool> IsLocked(int planogramId, CurrentUser user)
+        private async Task<bool> IsLocked(long planogramId, CurrentUser user)
         {
             var filter = new PlanogramLockFilter
             {
@@ -1297,12 +1297,12 @@ namespace PlanMatr_API.Controllers.planm
             }
         }
 
-        private async Task UpdateScratchPad(int planogramId, PlanmShelfInfoList scratchPad)
+        private async Task UpdateScratchPad(long planogramId, PlanmShelfInfoList scratchPad)
         {
             //Handle the planogram now
             var shelves = scratchPad.shelfInfos;
             var parts = scratchPad.partInfos;
-            int? sPadId = 0;
+            long? sPadId = 0;
             if (parts.Any() || shelves.Any())
             {
                 if (parts.Any())
